Input

+++ Engine/Source/Vector2.h
@@ -8,7 +8,7 @@ struct Vector2
 
 	Vector2() = default;
 	Vector2(float x, float y) { this -> x = x;this -> y = y; }
-	Vector2(int x, int y) { this->x = x; this->y = y; }
+	Vector2(int x, int y) { this->x = (float)x; this->y = (float)y; }
 
 
 	Vector2 operator + (Vector2 v) const { return Vector2{ x + v.x, y + v.y }; }
@@ -23,4 +23,6 @@ struct Vector2
 
 	float LengthSqr() { return (x * x) + (y * y); }
 	float Length() { return std::sqrt((x * x) + (y * y)); } //to get actual length
+
+
 };
